% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_sim_data.R
\name{cgr_sim}
\alias{cgr_sim}
\title{Simulate Observations From a Conditional Gaussian Regression}
\usage{
cgr_sim(
  g,
  lvls = NULL,
  hom = FALSE,
  ns = 10000,
  nc = 1,
  alpha_rate = 10,
  beta_rate = 10,
  sigma_rate = 10,
  cell_rate = 5,
  beta_hom = 10L,
  sigma_hom = 10L
)
}
\arguments{
\item{g}{Adjacency list of a decomposable (mixed) graph}

\item{lvls}{Levels of the discrete variables. A named list.}

\item{hom}{Logical indicating if the model is homogeneous or not}

\item{ns}{Number of simulations}

\item{nc}{Number of cores to simulate the discrete part}

\item{alpha_rate}{Control for intercept on cell level}

\item{beta_rate}{Control for slopes on cell level}

\item{sigma_rate}{Control for variances on cell level}

\item{cell_rate}{Control discrete cell probabilities}

\item{beta_hom}{Control for fixed slope if \code{hom = TRUE}}

\item{sigma_hom}{Control for fixed variance if \code{hom = TRUE}}
}
\description{
Simulate Observations From a Conditional Gaussian Regression
}
\details{
cell_rate, beta_hom and sigma_hom are not used in the pure_cont case
and when lvls is NULL it is assumed that g is of type pure_cont
}
