% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_graphd.R
\name{fit_mixed_graph}
\alias{fit_mixed_graph}
\title{Fit a marked decomposable graphical model}
\usage{
fit_mixed_graph(d, tree = FALSE, hom = FALSE, st = "BIC")
}
\arguments{
\item{d}{data.frame}

\item{tree}{Logical indicating whether or not the function should fit a tree or a general graph}

\item{hom}{Logical.}

\item{st}{Character. One of LR, AIC, or BIC}
}
\value{
An adjacency list
}
\description{
A generic method for structure learning in marked decomposable graphical models
}
\examples{
\dontrun{
# TBA
}
}
