% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fit_outlier.R
\name{fit_outlier_model}
\alias{fit_outlier_model}
\title{Outlier model}
\usage{
fit_outlier_model(
  A,
  adj,
  nsim = 10000,
  ncores = 1,
  validate = TRUE,
  approx = TRUE,
  hom = FALSE
)
}
\arguments{
\item{A}{data,frame with the new observation of interest appended}

\item{adj}{Adjacency list or gengraph object of a decomposable graph without the
observation of interest. See details}

\item{nsim}{Number of simulations}

\item{ncores}{Number of cores to use in parallelization}

\item{validate}{Logical. See details.}

\item{approx}{Logical. The user should rarely set this manually.}

\item{hom}{Logical indicating wheter the model is homogeneous or not if \code{A} has mixed variables}
}
\description{
A model based on decomposable graphical models for outlier detection
}
\details{
It is assumed that all cell values in \code{A}, for all variables,
are represented as a single character. If \code{validate} is \code{TRUE} this is checked.
If cell values are not single characters, one may exploit the \code{to_single_chars} function.

The \code{adj} object is most typically found using \code{fit_graph}. But the user can supply
an ajacency list, just a named \code{list}, of their own choice if needed.

The parameter \code{approx} is used to output the correct type of discrete simulations
whether or not an approximation is to be used. \code{fit_outlier} is an exact test
and calls \code{fit_outlier_model} with \code{approx = TRUE}.
}
