% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_decomp_graphs.R
\name{is_decomposable}
\alias{is_decomposable}
\title{A test for decomposability in undirected (possibly marked) graphs}
\usage{
is_decomposable(adj, disc_vars = NULL)
}
\arguments{
\item{adj}{Adjacency list of an undirected graph}

\item{disc_vars}{Character vector of discrete variables in adj if the graph is marked}
}
\description{
This function returns \code{TRUE} if the graph is decomposable and \code{FALSE} otherwise
}
\examples{
# Non-marked graphs:
# 4-cycle:
adj1 <- list(a = c("b", "d"), b = c("a", "c"), c = c("b", "d"), d = c("a", "c"))
is_decomposable(adj1) # FALSE
# Two triangles:
adj2 <- list(a = c("b", "d"), b = c("a", "c", "d"), c = c("b", "d"), d = c("a", "c", "b"))
is_decomposable(adj2) # TRUE

# Marked graphs:
# A line graph with a forbidden path: the definition of non-decomposability in marked graphs
adj3 <- list(a = "b", b = c("a", "c"), c = "b")
is_decomposable(adj, c("a", "c"))
}
