% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcs_marked}
\alias{mcs_marked}
\title{Maximum Cardinality Search in Marked Graphs}
\usage{
mcs_marked(adj, disc_vars, check = TRUE)
}
\arguments{
\item{adj}{A named adjacency list of a decomposable marked graph}

\item{disc_vars}{Character vector of discrete variables}

\item{check}{Boolean: check if adj is decomposable}
}
\value{
A list with a perfect numbering of the nodes and a perfect sequence of sets
}
\description{
Maximum Cardinality Search in Marked Graphs
}
\details{
If adj is not the adjacency list of a decomposable marked graph an error is raised
}
\examples{
x <- list(a = c("b", "d"), b = c("a", "c", "d"), c = c("b", "d"), d = c("a", "c", "b"))
mcs_marked(x, c("a"))
}
