% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odmgm.R
\docType{package}
\name{odmgm-package}
\alias{odmgm}
\alias{odmgm-package}
\title{odmgm: Outlier Detection in Mixed Graphical Models}
\description{
Contains functions to generate the plots and statistical
	     outputfound in the paper.
}
\author{
\strong{Maintainer}: Mads Lindskou \email{mads@math.aau.dk}

}
