% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{plot.gengraph}
\alias{plot.gengraph}
\title{Plot}
\usage{
\method{plot}{gengraph}(x, vc = NULL, ...)
}
\arguments{
\item{x}{A \code{gengraph} object}

\item{vc}{Named character vector; the names are the vertices and
the elements are the colors of the nodes}

\item{...}{Extra arguments. See the igraph package}
}
\description{
A wrapper around igraphs plot method for \code{gengraph} objects
}
