% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_decomp_graphs.R
\name{pmf}
\alias{pmf}
\title{Probability mass function for a decomposable graphical model}
\usage{
pmf(A, adj, logp = FALSE)
}
\arguments{
\item{A}{A character matrix of data}

\item{adj}{Adjacency list or gengraph object of a decomposable graph. See \code{fit_graph}.}

\item{logp}{Logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A function - the probability mass function corresponding
to the decomposable graph \code{adj} using \code{A} to estimate
the probabilities.
}
\description{
Probability mass function for a decomposable graphical model
}
\details{
It should be noted, that \code{pmf} is a closure function; i.e. its return
value is a function. Once \code{pmf} is called, one can query probabilities from the
returned function. If the probability of an observation is zero, the function will return
\code{NA} if \code{logp} is set to \code{TRUE} and zero otherwise.
}
