% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_outlier.R
\name{query2}
\alias{query2}
\alias{query2.outlier_model}
\title{Query Outlier Tests - version 2}
\usage{
query2(m, z, alpha = 0.05)

\method{query2}{outlier_model}(m, z, alpha = 0.05)
}
\arguments{
\item{m}{A \code{outlier_model} object}

\item{z}{data.frame with new data}

\item{alpha}{The significance level}
}
\description{
Given an outlier model, an outlier test for each row in the new data is conducted
}
\details{
Use \code{fit_outlier_model} to obtain \code{m}
}
