% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rip}
\alias{rip}
\title{Running Intersection Property}
\usage{
rip(adj, check = TRUE)
}
\arguments{
\item{adj}{A named adjacency list of a decomposable graph}

\item{check}{Boolean: check if adj is decomposable}
}
\description{
Given a decomposable graph, this functions finds a perfect numbering on the vertices using maximum cardinality search, and hereafter returns a list with two elements: "C" - the cliques and "S" - the separators.

Given a decomposable graph, this functions finds a perfect numbering on the vertices
using maximum cardinality search, and hereafter returns a list with three elements:
"C"   - the cliques,
"S"   - the separators and
"PA" - the parents (all vertices adjacent to the node i that are numbered before node i)
}
\examples{
x <- list(a = c("b", "d"), b = c("a", "c", "d"), c = c("b", "d"), d = c("a", "c", "b"))
y <- rip(x)
# Cliques:
y$C
# Separators:
y$S
# Parents:
y$PA
}
\seealso{
\code{\link{mcs}}, \code{\link{is_decomposable}}
}
