% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rip_marked}
\alias{rip_marked}
\title{Running Intersection Property for Marked Graphs}
\usage{
rip_marked(adj, disc_vars, check = TRUE)
}
\arguments{
\item{adj}{A named adjacency list of a decomposable marked graph}

\item{disc_vars}{Character vector of discrete variables in adj}

\item{check}{Boolean: check if adj is decomposable}
}
\description{
Given a decomposable marked graph, this functions finds a perfect numbering on the vertices
using maximum cardinality search, and hereafter returns a list with three elements:
"C"   - the cliques,
"S"   - the separators and
"PA" - the parents (all vertices adjacent to the node i that are numbered before node i)
}
\examples{
x <- list(a = c("b", "d"), b = c("a", "c", "d"), c = c("b", "d"), d = c("a", "c", "b"))
y <- rip_marked(x, c("a"))
# Cliques:
y$C
# Separators:
y$S
# Parents:
y$PA
}
\seealso{
\code{\link{mcs_marked}}, \code{\link{is_decomposable}}
}
