% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_misc.R
\name{to_single_chars}
\alias{to_single_chars}
\title{To Single Chars}
\usage{
to_single_chars(x)
}
\arguments{
\item{x}{Data frame or matrix of characters}
}
\description{
Convert all values in a data frame or matrix of characters to a single character representation
}
\examples{

## Example 1 - pure discrete
d1 <- data.frame(x = c("hhh", "f"), y = c("f", "hhh"), stringsAsFactors = FALSE)
to_single_chars(d1)

## Example 2 - mixed variables
library(dplyr)
d2 <- cbind(d1, data.frame(z = runif(2), w = runif(2)))

d2_disc <- d2 \%>\%
        select_if(is.character) \%>\%
        to_single_chars()

d2_cont <- d2 \%>\% select_if(is.numeric)

cbind(d2_disc, d2_cont)
}
